feat/main - ui viewer xml layout
