+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
+        <activity android:name=".ui.viewer.ViewerActivity"></activity>
         <activity
             android:name=".ui.main.MainActivity"
             android:windowSoftInputMode="stateHidden">
