+++ app/src/main/java/com/hackday/imageSearch/di/ViewModelModule.kt
@@ -1,8 +1,9 @@
 package com.hackday.imageSearch.di
 
-import com.hackday.imageSearch.ui.main.AlbumViewModel
+import com.hackday.imageSearch.ui.album.AlbumViewModel
 import com.hackday.imageSearch.ui.main.MainViewModel
-import com.hackday.imageSearch.ui.main.PhotoViewModel
+import com.hackday.imageSearch.ui.photo.PhotoViewModel
+import com.hackday.imageSearch.ui.viewer.ViewerViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
 
@@ -10,4 +11,5 @@ val viewModelModule = module {
     viewModel { MainViewModel() }
     viewModel { PhotoViewModel() }
     viewModel { AlbumViewModel() }
+    viewModel { ViewerViewModel() }
 }
