+++ app/src/main/java/com/hackday/imageSearch/ui/album/AlbumFragment.kt
@@ -1,4 +1,4 @@
-package com.hackday.imageSearch.ui.main
+package com.hackday.imageSearch.ui.album
 
 import android.os.Bundle
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -6,8 +6,8 @@ import androidx.recyclerview.widget.RecyclerView
 import com.hackday.imageSearch.R
 import com.hackday.imageSearch._base.BaseFragment
 import com.hackday.imageSearch.databinding.FragmentAlbumBinding
-import com.hackday.imageSearch.ui.main.adapterAlbum.AlbumItemDecorator
-import com.hackday.imageSearch.ui.main.adapterAlbum.AlbumRecyclerAdapter
+import com.hackday.imageSearch.ui.album.adapter.AlbumItemDecorator
+import com.hackday.imageSearch.ui.album.adapter.AlbumRecyclerAdapter
 import kotlinx.android.synthetic.main.fragment_album.*
 import org.koin.androidx.viewmodel.ext.android.viewModel
