+++ app/src/main/java/com/hackday/imageSearch/ui/album/adapter/AlbumItemDecorator.kt
@@ -1,12 +1,11 @@
-package com.hackday.imageSearch.ui.main.adapterPhoto
-
+package com.hackday.imageSearch.ui.album.adapter
 
 import android.graphics.Rect
 import android.util.TypedValue
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 
-class PhotoItemDecorator : RecyclerView.ItemDecoration() {
+class AlbumItemDecorator : RecyclerView.ItemDecoration() {
     override fun getItemOffsets(
         outRect: Rect,
         view: View,
@@ -17,20 +16,15 @@ class PhotoItemDecorator : RecyclerView.ItemDecoration() {
 
         val horizontalValue = TypedValue.applyDimension(
             TypedValue.COMPLEX_UNIT_DIP,
-            2f,
-            view.resources.displayMetrics
-        ).toInt()
-
-        val verticalValue = TypedValue.applyDimension(
-            TypedValue.COMPLEX_UNIT_DIP,
-            1f,
+            15f,
             view.resources.displayMetrics
         ).toInt()
 
 
-        when (parent.getChildLayoutPosition(view) % 4) {
-            3 -> outRect.set(0, 0, 0, verticalValue)
-            else -> outRect.set(0, 0, horizontalValue, verticalValue)
+        when (parent.getChildLayoutPosition(view)) {
+            0 -> outRect.set(horizontalValue, 0, horizontalValue, 0)
+            state.itemCount - 1 -> outRect.set(0, 0, horizontalValue, 0)
+            else -> outRect.set(0, 0, horizontalValue, 0)
         }
 
     }
