+++ app/src/main/java/com/hackday/imageSearch/ui/main/adapter/BindingAdapter.kt
@@ -1,8 +1,9 @@
-package com.hackday.imageSearch.ui.main.adapterAlbum
+package com.hackday.imageSearch.ui.main.adapter
 
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.hackday.imageSearch.ui.main.adapterPhoto.PhotoRecyclerAdapter
+import com.hackday.imageSearch.ui.album.adapter.AlbumRecyclerAdapter
+import com.hackday.imageSearch.ui.photo.adapter.PhotoRecyclerAdapter
 
 class BindingAdapter {
     companion object {
