+++ app/src/main/java/com/hackday/imageSearch/ui/photo/adapter/PhotoItemDecorator.kt
@@ -1,11 +1,12 @@
-package com.hackday.imageSearch.ui.main.adapterAlbum
+package com.hackday.imageSearch.ui.photo.adapter
+
 
 import android.graphics.Rect
 import android.util.TypedValue
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
 
-class AlbumItemDecorator : RecyclerView.ItemDecoration() {
+class PhotoItemDecorator : RecyclerView.ItemDecoration() {
     override fun getItemOffsets(
         outRect: Rect,
         view: View,
@@ -16,15 +17,20 @@ class AlbumItemDecorator : RecyclerView.ItemDecoration() {
 
         val horizontalValue = TypedValue.applyDimension(
             TypedValue.COMPLEX_UNIT_DIP,
-            15f,
+            2f,
+            view.resources.displayMetrics
+        ).toInt()
+
+        val verticalValue = TypedValue.applyDimension(
+            TypedValue.COMPLEX_UNIT_DIP,
+            1f,
             view.resources.displayMetrics
         ).toInt()
 
 
-        when (parent.getChildLayoutPosition(view)) {
-            0 -> outRect.set(horizontalValue, 0, horizontalValue, 0)
-            state.itemCount - 1 -> outRect.set(0, 0, horizontalValue, 0)
-            else -> outRect.set(0, 0, horizontalValue, 0)
+        when (parent.getChildLayoutPosition(view) % 4) {
+            3 -> outRect.set(0, 0, 0, verticalValue)
+            else -> outRect.set(0, 0, horizontalValue, verticalValue)
         }
 
     }
