+++ app/src/main/java/com/hackday/imageSearch/ui/viewer/ViewerActivity.kt
@@ -1,0 +1,21 @@
+package com.hackday.imageSearch.ui.viewer
+
+import android.os.Bundle
+import com.hackday.imageSearch.R
+import com.hackday.imageSearch._base.BaseActivity
+import com.hackday.imageSearch.databinding.ActivityViewerBinding
+import org.koin.androidx.viewmodel.ext.android.viewModel
+
+class ViewerActivity : BaseActivity<ActivityViewerBinding>() {
+    override val vm: ViewerViewModel by viewModel()
+    override fun getLayoutRes() = R.layout.activity_viewer
+
+    override fun setupBinding() {
+        binding.vm = vm
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+    }
+}
+
