+++ app/src/main/java/com/hackday/imageSearch/ui/viewer/ViewerViewModel.kt
@@ -1,0 +1,7 @@
+package com.hackday.imageSearch.ui.viewer
+
+import androidx.lifecycle.ViewModel
+
+class ViewerViewModel : ViewModel() {
+
+}
