+++ app/src/main/res/layout/fragment_album.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="vm"
-            type="com.hackday.imageSearch.ui.main.AlbumViewModel" />
+            type="com.hackday.imageSearch.ui.album.AlbumViewModel" />
 
     </data>
